Fix FS#10476. Prevent FLAC bitrate calculation overflow with large files. The watermark depends on the bitrate, and so this also fixes playback pauses on large FLAC files.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@22211 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/flac.c
@@ -172,7 +172,8 @@ static bool flac_init(FLACContext* fc, int first_frame_offset)
     }
 
    if (found_streaminfo) {
-       fc->bitrate = ((fc->filesize-fc->metadatalength) * 8) / fc->length;
+       fc->bitrate = ((int64_t) (fc->filesize-fc->metadatalength) * 8) 
+                     / fc->length;
        return true;
    } else {
        return false;
